Automated Daily Inventory Reporting/requirements.txt
pandas
google-api-python-client
google-auth
python-dotenv